requirements.txt
boto3==1.16.52
botocore==1.19.52
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
django-storages==1.11.1
ipython==7.16.1
# pandjas:
git+https://github.com/StornWhite/pandjas.git@v0.1.2
psycopg2-binary==2.8.3
